refactor(utils): move search formatting logic to FormatUtil
- Extract EmbedFactory and formatting logic out of `AudioLoadResultHandler.java` into `FormatUtil.java`
